Revert "Modificacion tipo de dato precios a LONG DOUBLE"
This reverts commit 3ab10470843d7a00f41f561c7995c58c38274019.

+++ proyecto-codeblocks-dev/include/Producto.h
@@ -11,7 +11,7 @@ class Producto
 public:
     //CONSTRUCTORES
     Producto();
-    Producto(int ID, int categoria, int idMarca, std::string modelo, std::string descripcion, long double precioVenta, int stock, bool activo);
+    Producto(int ID, int categoria, int idMarca, std::string modelo, std::string descripcion, float precioVenta, int stock, bool activo);
 
     //SETTERS
     void setID(int ID);
@@ -19,7 +19,7 @@ public:
     void setMarca (int idMarca);
     void setModelo(std::string modelo);
     void setDescripcion(std::string descripcion);
-    void setPrecio (long double precioVenta);
+    void setPrecio (float precioVenta);
     void setStock (int stock);
     void setActivo (bool activo);
 
@@ -29,7 +29,7 @@ public:
     int getIdMarca ();
     std::string getModelo();
     std::string getDescripcion();
-    long double getPrecio ();
+    float getPrecio ();
     int getStock ();
     bool getActivo ();
 
@@ -39,7 +39,7 @@ private:
     int _idMarca;
     char _modelo[15];
     char _descripcion[50];
-    long double _precioVenta;
+    float _precioVenta;
     int _stock;
     bool _activo;
 };
